admin 로직 구현

+++ src/main/java/com/c1/donguri/template/TemplateDeleteAdminC.java
@@ -1,5 +1,8 @@
 package com.c1.donguri.template;
 
+import com.c1.donguri.user.UserDAO;
+import com.c1.donguri.user.UserDTO;
+
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -11,7 +14,20 @@ import java.io.IOException;
 public class TemplateDeleteAdminC extends HttpServlet {
 
     public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
+        boolean isLoggedIn = UserDAO.USER_DAO.loginCheck(request);
 
+        if (!isLoggedIn) {
+            response.sendRedirect("login");
+            return;
+        }
+        UserDTO user = (UserDTO) request.getSession().getAttribute("user");
+
+        boolean isAdmin = UserDAO.USER_DAO.isAdmin(user.getEmail());
+
+        if (!isAdmin) {
+            response.sendRedirect("main");
+            return;
+        }
         TemplateDAO.TEMPLATE_DAO.deleteTemplate(request);
 
         response.sendRedirect("/template-list-admin");
